+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* Add an `--output-file` CLI argument to make the executable write to a file
+  instead of print to stdout.
 * Use the `cmdargs` package to parse CLI arguments.
 * Split library into `Console.BnbStaking` module hierarchy.